pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "clickhouse-connect",
    "schedule",
    "pydantic-settings>=2.10.1",
    "supabase>=2.16.0",
    "postgrest>=1.1.1",
    "pandas-stubs==2.2.3.250527",
    "python-dotenv>=0.21.0",
]

[dependency-groups]
dev = [
    "dtale>=3.17.0",
    "itables>=2.4.2",
    "kagglehub>=0.3.12",
]
